+++ apps/blog/src/server/libs/sudoku/enum.ts
@@ -1,0 +1,6 @@
+export enum SudokuDifficulty {
+  Beginner = "beginner",
+  Medium = "medium",
+  Hard = "hard",
+  Expert = "expert",
+}