+++ jest.config.js
@@ -1,0 +1,15 @@
+module.exports = {
+  roots: ["<rootDir>"],
+  moduleFileExtensions: ["ts", "tsx", "js", "json", "jsx"],
+  testEnvironment: "jsdom",
+  testPathIgnorePatterns: ["<rootDir>[/\\\\](node_modules|.next)[/\\\\]"],
+  transformIgnorePatterns: ["[/\\\\]node_modules[/\\\\].+\\.(ts|tsx)$"],
+  transform: {
+    "^.+\\.(ts|tsx)$": "babel-jest",
+  },
+  watchPlugins: ["jest-watch-typeahead/filename", "jest-watch-typeahead/testname"],
+  moduleNameMapper: {
+    "^@/(.*)$": "<rootDir>/src/$1",
+    "^@howardism/(.*)$": "<rootDir>/../../packages/$1",
+  },
+};